+++ src/br/gov/cesarschool/poo/bonusvendas/entidade/LancamentoBonusCredito.java
@@ -4,7 +4,7 @@ package br.gov.cesarschool.poo.bonusvendas.entidade;
 import java.time.LocalDateTime;
 
 public class LancamentoBonusCredito extends LancamentoBonus{
-    public LancamentoBonusCredito(long numeroCaixaDeBonus, int valor, LocalDateTime dataHoraLancamento) {
+    public LancamentoBonusCredito(long numeroCaixaDeBonus, double valor, LocalDateTime dataHoraLancamento) {
         super(numeroCaixaDeBonus, valor, dataHoraLancamento);
     }
 }
